requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
python-docx>=1.1.0
openai>=1.30.0
anthropic>=0.28.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
